Refactor test helpers to their own package.
