+++ pkg/test/clients/clients.go
@@ -1,17 +1,9 @@
-package test
+package clients
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"os"
 	"testing"
 
-	"github.com/google/go-github/v34/github"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/apis/pipelinesascode/v1alpha1"
 	fakepacclientset "github.com/openshift-pipelines/pipelines-as-code/pkg/generated/clientset/versioned/fake"
 	informersv1alpha1 "github.com/openshift-pipelines/pipelines-as-code/pkg/generated/informers/externalversions/pipelinesascode/v1alpha1"
@@ -44,45 +36,6 @@ type Data struct {
 	PipelineRuns []*pipelinev1alpha1.PipelineRun
 	Repositories []*v1alpha1.Repository
 	Namespaces   []*corev1.Namespace
-}
-
-const (
-	// baseURLPath is a non-empty Client.BaseURL path to use during tests,
-	// to ensure relative URLs are used for all endpoints. See issue #752.
-	githubBaseURLPath = "/api-v3"
-)
-
-// SetupGH Setup a GitHUB httptest connexion, from go-github test-suit
-func SetupGH() (client *github.Client, mux *http.ServeMux, serverURL string, teardown func()) {
-	// mux is the HTTP request multiplexer used with the test server.
-	mux = http.NewServeMux()
-
-	// We want to ensure that tests catch mistakes where the endpoint URL is
-	// specified as absolute rather than relative. It only makes a difference
-	// when there's a non-empty base URL path. So, use that. See issue #752.
-	apiHandler := http.NewServeMux()
-	apiHandler.Handle(githubBaseURLPath+"/", http.StripPrefix(githubBaseURLPath, mux))
-	apiHandler.HandleFunc("/", func(w http.ResponseWriter, req *http.Request) {
-		fmt.Fprintln(os.Stderr, "FAIL: Client.BaseURL path prefix is not preserved in the request URL:")
-		fmt.Fprintln(os.Stderr)
-		fmt.Fprintln(os.Stderr, "\t"+req.URL.String())
-		fmt.Fprintln(os.Stderr)
-		fmt.Fprintln(os.Stderr, "\tDid you accidentally use an absolute endpoint URL rather than relative?")
-		fmt.Fprintln(os.Stderr, "\tSee https://github.com/google/go-github/issues/752 for information.")
-		http.Error(w, "Client.BaseURL path prefix is not preserved in the request URL.", http.StatusInternalServerError)
-	})
-
-	// server is a test HTTP server used to provide mock API responses.
-	server := httptest.NewServer(apiHandler)
-
-	// client is the GitHub client being tested and is
-	// configured to use test server.
-	client = github.NewClient(nil)
-	url, _ := url.Parse(server.URL + githubBaseURLPath + "/")
-	client.BaseURL = url
-	client.UploadURL = url
-
-	return client, mux, server.URL, server.Close
 }
 
 // SeedTestData returns Clients and Informers populated with the
@@ -120,33 +73,4 @@ func SeedTestData(t *testing.T, ctx context.Context, d Data) (Clients, Informers
 
 	c.PipelineAsCode.ClearActions()
 	return c, i
-}
-
-// roundTripFunc .
-type roundTripFunc func(req *http.Request) *http.Response
-
-// RoundTrip .
-func (f roundTripFunc) RoundTrip(req *http.Request) (*http.Response, error) {
-	return f(req), nil
-}
-
-// NewTestClient returns *http.Client with Transport replaced to avoid making real calls
-func newHTTPTestClient(fn roundTripFunc) *http.Client {
-	return &http.Client{
-		Transport: roundTripFunc(fn),
-	}
-}
-
-func MakeHTTPTestClient(t *testing.T, statusCode int, body string) *http.Client {
-	httpTestClient := newHTTPTestClient(func(req *http.Request) *http.Response {
-		// Test request parameters
-		return &http.Response{
-			StatusCode: statusCode,
-			// Send response to be tested
-			Body: ioutil.NopCloser(bytes.NewBufferString(body)),
-			// Must be set to non-nil value or it panics
-			Header: make(http.Header),
-		}
-	})
-	return httpTestClient
 }
