+++ pkg/test/kubernetestint/kubernetesint.go
@@ -1,4 +1,4 @@
-package test
+package kubernetestint
 
 import "errors"
